requirements.txt
asgiref==3.2.3
celery==4.4.0
Django==2.2
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
djangorestframework==3.11.0
python-dotenv==0.10.5
